fix bug

+++ database/migrations/2024_08_10_063547_create_label_imports_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('label');
             $table->string('sheet_name')->nullable();
             $table->string('cell')->nullable();
-            $table->float('nilai')->nullable();
+            $table->decimal('nilai', 20, 2)->nullable();
             $table->timestamps();
         });
     }
